requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
jinja2==3.1.6
requests==2.32.4
python-dotenv==1.1.1
pydantic==2.11.7
openpyxl==3.1.5
pandas==2.3.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
